% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluestimate.R
\name{Cluestimate}
\alias{Cluestimate}
\title{Estimate the number of cell clusters in a single-cell dataset}
\usage{
Cluestimate(procdata, K_max = 20)
}
\arguments{
\item{procdata}{The preprocessed single cell data, which is a numeric matrix or data.frame. Rows represent genes/features cluster and columns represent single cells.}

\item{K_max}{Maximum number of clusters. If NULL , then it will be set as 20.}
}
\value{
The optimal number of clustering clusters.
}
\description{
Estimate the number of cell clusters in a single-cell dataset
}
\examples{
data(lpsdata)
procdata = preprocess(lpsdata, takenormalize=TRUE, clusternum=350)
K_opt = Cluestimate(procdata, K_max=20)
}
