% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOA_measure_convex.R
\name{FOA_measure_convex}
\alias{FOA_measure_convex}
\title{Estimate vertices of convex sets}
\usage{
FOA_measure_convex(X, N)
}
\arguments{
\item{X}{Convex set, which is a numeric matrix or data.frame. Rows represent genes/features cluster and columns represent single cells.}

\item{N}{The number of vertices}
}
\value{
Estimated vertices number
}
\description{
Estimate vertices of convex sets
}
\examples{
Y=FOA_measure_convex(X,N)
}
