% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCAM.R
\name{FSCAM}
\alias{FSCAM}
\title{FSCAM: CAM-based feature selection for clustering scRNA-seq}
\usage{
FSCAM(procdata, K)
}
\arguments{
\item{procdata}{The preprocessed single cell data, which is a numeric matrix or data.frame. Rows represent genes/features cluster centers and columns represent single cells.}

\item{K}{The number of cell clusters. If NULL , then it will be estimated by function Cluestimate.}
}
\value{
The name vector of the selected DE genes.
}
\description{
FSCAM: CAM-based feature selection for clustering scRNA-seq
}
\examples{
data(lpsdata)
procdata = preprocess(lpsdata, takenormalize=TRUE, clusternum=350)
DEGs =FSCAM(procdata,K)
}
