% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnls2.R
\name{nnls2}
\alias{nnls2}
\title{Title}
\usage{
nnls2(C, d)
}
\arguments{
\item{C}{The independent variables}

\item{d}{The dependent variable}
}
\value{
the Regression coefficient vector x that minimizes Norm(d-C*x)
}
\description{
Title
}
\examples{
x=nnls(C,d)
}
