% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(data, takenormalize = TRUE, clusternum = 350)
}
\arguments{
\item{data}{The raw single_cell data, which is a numeric matrix or data.frame. Rows represent genes/features and columns represent single cells.}

\item{takenormalize}{Logical value indicating whether to take normalize. If FALSE no normalizing will be performed.}

\item{clusternum}{The number of clusters for doing cluster, typically 5% of number of all genes. The clustering will be done after all the gene filtering and normalizing. If NULL no clustering will be performed.}
}
\value{
Preprocessed cluster-level expression matrix and cluster to which each gene belongs
}
\description{
preprocess
}
\examples{
data(lpsdata)
procdata = preprocess(lpsdata, takenormalize=TRUE, clusternum=350)
}
