% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCC_FSCAM.R
\name{SCC_FSCAM}
\alias{SCC_FSCAM}
\title{Single-cell clustering algorithm based on FSCAM method}
\usage{
SCC_FSCAM(procdata, K = NULL, T = 20)
}
\arguments{
\item{procdata}{The preprocessed single cell data, which is a numeric matrix or data.frame. Rows represent genes/features cluster and columns represent single cells.}

\item{K}{The number of cell clusters. If NULL , then it will be estimated by function Cluestimate.}

\item{T}{Running times of FSCAM feature selection algorithm. If NULL , then it will be set as 20.}
}
\value{
The name vector of the selected features and K types of cell-labels for cells in input dataset.
}
\description{
Single-cell clustering algorithm based on FSCAM method
}
\examples{
data(lpsdata)
procdata = preprocess(lpsdata, takenormalize=TRUE, clusternum=350)
DEGs = FSCAM(procdata,K)
features_labels = SCC_FSCAM(procdata,K=NULL, T=20)
}
